Fix:#25 use separate database and allow connection to external database
