--- app/config.py
@@ -1,17 +1,0 @@
-
-class Config:
-    SECRET_KEY = "your-secret-key-here"
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-    UPLOAD_FOLDER = 'uploads'
-    MAX_CONTENT_LENGTH = 5 * 1024 * 1024  # 5MB
-
-class DevelopmentConfig(Config):
-    DEBUG = True
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///dev.db'
-
-class TestingConfig(Config):
-    TESTING = True
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///test.db'
-
-class ProductionConfig(Config):
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///prod.db'
