+++ main.py
@@ -2,7 +2,12 @@
 
 import os
 from app import create_app, db
-from app.config import DevelopmentConfig
+from config import DevelopmentConfig
+
+from dotenv import load_dotenv
+
+load_dotenv()
+print("Loaded DATABASE_URL:", os.getenv("DATABASE_URL"))
 
 app = create_app(DevelopmentConfig)
